pyproject.toml
[project]
name = "chatbot-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.26.0",
    "langchain>=0.3.26",
    "load-dotenv>=0.1.0",
    "openai>=1.97.0",
]
